GUI refactor and modularization
