--- src/main/java/algoblocks/gui/ActionBlockItem.java
@@ -1,13 +1,0 @@
-package algoblocks.gui;
-
-import algoblocks.engine.block.*;
-
-public class ActionBlockItem extends BlockItem {
-    public ActionBlockItem(Block block) {
-        super(block);
-    }
-
-    public String toString() {
-        return block.getName();
-    }
-}
