--- src/main/java/algoblocks/gui/BlockItem.java
@@ -1,14 +1,0 @@
-package algoblocks.gui;
-
-import algoblocks.engine.block.*;
-
-public abstract class BlockItem {
-    protected Block block;
-
-    public BlockItem(Block block) {
-        this.block = block;
-    }
-
-    @Override
-    public abstract String toString();
-}
