--- src/main/java/algoblocks/gui/EffectBlockItemClosedState.java
@@ -1,13 +1,0 @@
-package algoblocks.gui;
-
-public class EffectBlockItemClosedState extends EffectBlockItemState {
-    @Override
-    public String toString() {
-        return "Closed";
-    }
-    
-    @Override
-    public boolean isOpened() {
-        return false;
-    }
-}
