--- src/main/java/algoblocks/gui/EffectBlockItemOpenedState.java
@@ -1,13 +1,0 @@
-package algoblocks.gui;
-
-public class EffectBlockItemOpenedState extends EffectBlockItemState {
-    @Override
-    public String toString() {
-        return "Opened";
-    }
-
-    @Override
-    public boolean isOpened() {
-        return true;
-    }
-}
