--- src/main/java/algoblocks/gui/EffectBlockItemState.java
@@ -1,6 +1,0 @@
-package algoblocks.gui;
-
-public abstract class EffectBlockItemState {
-    public abstract String toString();
-    public abstract boolean isOpened();
-}
