--- src/main/java/algoblocks/gui/SpritesPath.java
@@ -1,11 +1,0 @@
-package algoblocks.gui;
-
-
-public class SpritesPath{
-    public static final String RUN_PENCIL_DOWN = "File:resources/runPencilDown.png";
-    public static final String RUN_PENCIL_UP = "File:resources/runPencilUp.png";
-    public static final String IDLE_PENCIL_DOWN = "File:resources/idlePencilDown.png";
-    public static final String IDLE_PENCIL_UP = "File:resources/idlePencilUp.png";
-
-    private SpritesPath(){}
-}
