--- src/main/java/algoblocks/gui/buttons/AlgorithmButtonNames.java
@@ -1,12 +1,0 @@
-package algoblocks.gui.buttons;
-
-
-public class AlgorithmButtonNames{
-    public static final String CLOSE_BLOCK_BUTTON = "Close block";
-    public static final String DELETE_ALL_BUTTON = "Delete All";
-    public static final String DELETE_BUTTON = "Delete";
-    public static final String RUN_BUTTON = "Run";
-    public static final String SAVE_BUTTON = "Save";
-
-    private AlgorithmButtonNames(){}
-}
