+++ src/main/java/algoblocks/view/buttons/RunButton.java
@@ -1,0 +1,25 @@
+package algoblocks.view.buttons;
+
+import algoblocks.view.*;
+
+import java.util.ArrayList;
+
+import javafx.scene.control.Button;
+import javafx.event.EventHandler;
+import javafx.event.ActionEvent;
+
+
+public class RunButton extends Button {
+    ArrayList<Button> buttonsToDisable;
+
+    public RunButton(RunAlgorithm run) {
+        super();
+        this.setText(AlgorithmButtonNames.RUN_BUTTON);
+        this.setOnAction(new EventHandler<ActionEvent>() {
+            @Override
+            public void handle(ActionEvent actionEvent) {
+                run.execute();
+            }
+        });
+    }
+}
