+++ src/main/java/algoblocks/view/whiteboard/WhiteboardCanvas.java
@@ -1,0 +1,33 @@
+package algoblocks.view.whiteboard;
+
+import algoblocks.controller.GameController;
+import algoblocks.view.whiteboard.player.PlayerPane;
+import javafx.scene.layout.StackPane;
+
+import java.util.ArrayList;
+
+public class WhiteboardCanvas {
+    private WhiteboardGrid grid;
+    private WhiteboardDrawing drawing;
+    private PlayerPane playerPane;
+
+    public WhiteboardCanvas(GameController game, double width, double height) {
+        grid = new WhiteboardGrid(game, width, height);
+        drawing = new WhiteboardDrawing(game, width, height);
+        playerPane = new PlayerPane(game, width, height);
+    }
+
+    public StackPane getStackPane() {
+        StackPane stackPane = new StackPane();
+        stackPane.getChildren().addAll(grid.getPane(), drawing.getPane(), playerPane.getPane());
+        return stackPane;
+    }
+
+    public ArrayList<Runnable> draw() {
+        return drawing.draw(playerPane);
+    }
+
+    public void clear() {
+        drawing.clear();
+    }
+}
