+++ src/main/java/algoblocks/view/whiteboard/player/PlayerSprite.java
@@ -1,0 +1,19 @@
+ package algoblocks.view.whiteboard.player;
+
+ import javafx.scene.image.Image;
+
+public class PlayerSprite{
+    public PlayerSprite(){
+
+    }
+
+    public Image get(boolean canDraw, boolean isMoving){
+        if (isMoving) {
+            if (canDraw) return new Image(SpritesPath.RUN_PENCIL_DOWN);
+            return new Image(SpritesPath.RUN_PENCIL_UP);
+        } else {
+            if (canDraw) return new Image(SpritesPath.IDLE_PENCIL_DOWN);
+            return new Image(SpritesPath.IDLE_PENCIL_UP);
+        }
+    }
+}
