+++ src/main/java/algoblocks/view/workspace/block/ActionBlockItem.java
@@ -1,0 +1,13 @@
+package algoblocks.view.workspace.block;
+
+import algoblocks.engine.block.*;
+
+public class ActionBlockItem extends BlockItem {
+    public ActionBlockItem(Block block) {
+        super(block);
+    }
+
+    public String toString() {
+        return block.getName();
+    }
+}
