+++ src/main/java/algoblocks/view/workspace/block/BlockItem.java
@@ -1,0 +1,14 @@
+package algoblocks.view.workspace.block;
+
+import algoblocks.engine.block.*;
+
+public abstract class BlockItem {
+    protected Block block;
+
+    public BlockItem(Block block) {
+        this.block = block;
+    }
+
+    @Override
+    public abstract String toString();
+}
