+++ src/main/java/algoblocks/view/workspace/block/EffectBlockItemClosedState.java
@@ -1,0 +1,13 @@
+package algoblocks.view.workspace.block;
+
+public class EffectBlockItemClosedState extends EffectBlockItemState {
+    @Override
+    public String toString() {
+        return "Closed";
+    }
+    
+    @Override
+    public boolean isOpened() {
+        return false;
+    }
+}
