+++ src/main/java/algoblocks/view/workspace/block/EffectBlockItemOpenedState.java
@@ -1,0 +1,13 @@
+package algoblocks.view.workspace.block;
+
+public class EffectBlockItemOpenedState extends EffectBlockItemState {
+    @Override
+    public String toString() {
+        return "Opened";
+    }
+
+    @Override
+    public boolean isOpened() {
+        return true;
+    }
+}
