+++ src/main/java/algoblocks/view/workspace/block/EffectBlockItemState.java
@@ -1,0 +1,6 @@
+package algoblocks.view.workspace.block;
+
+public abstract class EffectBlockItemState {
+    public abstract String toString();
+    public abstract boolean isOpened();
+}
